Refactor app authentication and route layout to improve UX and clarity of code
This now follows a more straight forward pattern. On entering the app the user is redirected to login page if they are not logged in. They are then asked to login and redirected to the dashboard. The routes are all in one file and we now have an authenticated route component which checks authentication and renders the component that the user is heading towards.